Part1 finds corners in Image
Part 1 now finds corners of each card in a given image and returns the results as a list, where each row corresponds to a card and every two values within that list correspond to a corner. Also wrote a test program to demonstrate how to use the data given from Part1.