Reformatted Part 1 and Card Info Storage
Reformatted Part 1 into two files, IsolateCards and NormalizeCards, which will get the card bounds for each card in the original image and then convert the card to standard size (700 pixels x 500 pixels). Cards now stored in an array called cards, where each entry is a CardInfo object. CardInfo objects hold the bounds of the cards in reference to the original image and the normalized card by default, and the other 4 properties of the card (shape, pattern, color, number) will default to 0 / empty value, but can be set using a setter method. All properties of the CardInfo object have getter methods for acquiring values of said card. Also lots of warnings show up but no problems occur, so just a heads up with that.