Shape, Pattern and Number Detection
Added functions for detecting the shape, pattern, and number of a shape. Ran multiple tests and so far everything works really well. Only known issue is that in very dark lined images the pattern may be detected as empty rather than lined. Looking into way to work around this, but this is a very specific issue that rarely happens and might be worth just letting go.